+++ examples/print_ascii.c
@@ -6,13 +6,31 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/03 11:25:49 by asoursou          #+#    #+#             */
-/*   Updated: 2020/06/03 12:40:50 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/09/17 19:46:03 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-int	main(void)
+static void	print_char(int c)
+{
+	int p;
+
+	if (ft_isprint(c))
+	{
+		if (c == '\'')
+			ft_putstr("'\\''");
+		else
+			ft_printf("'%c'", c);
+	}
+	else
+	{
+		p = ft_toprint(c);
+		ft_printf((c == p ? "'\\x%hhx'" : "'\\%c'"), p);
+	}
+}
+
+int			main(void)
 {
 	int		i;
 	char	c;
@@ -21,7 +39,7 @@ int	main(void)
 	while (++i < 128)
 	{
 		c = i;
-		ft_print_char(&c);
+		print_char(c);
 		ft_putchar(i % 8 == 7 ? '\n' : '\t');
 	}
 	return (0);