+++ examples/rbtree.c
@@ -6,16 +6,15 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/02 20:58:11 by asoursou          #+#    #+#             */
-/*   Updated: 2020/08/26 14:29:08 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/09/17 20:09:04 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include <libft.h>
 
 #define CONTINUE "Press ENTER to continue..."
 
-static char		*get_line(char *label)
+static char		*get_line(const char *label)
 {
 	char *line;
 
@@ -24,42 +23,34 @@ static char		*get_line(char *label)
 	return (line);
 }
 
-static t_rbtree	*new_value(int n)
+static void		print_value(const void *content)
 {
-	t_rbtree	*elem;
-	int			*value;
-
-	if (!(value = malloc(sizeof(int))))
-		return (NULL);
-	if (!(elem = ft_rbtree_new(value)))
-		return (ft_memdel(value));
-	*value = n;
-	return (elem);
+	ft_printf("%lu", (t_u64)content);
 }
 
 static void		edit_tree(t_rbtree **tree, char op)
 {
 	char		*s;
-	int			n;
+	void		*n;
 	t_rbtree	*elem;
 	bool		confirm;
 
 	s = get_line("Enter a value: ");
-	n = s ? ft_atoi(s) : 0;
+	n = (void *)(s ? ft_strtoul(s, NULL, 10) : 0);
 	ft_memdel(s);
-	elem = ft_rbtree_search(*tree, &n, &ft_compare_int);
+	elem = ft_rbtree_search(*tree, n, &ft_compare_pointer);
 	confirm = true;
 	if (op == 'a')
 	{
 		if (elem)
 			ft_putendl_fd("error: This value already exists!", 2);
-		else if (!ft_rbtree_insert(tree, new_value(n), &ft_compare_int))
+		else if (!ft_rbtree_insert(tree, ft_rbtree_new(n), &ft_compare_pointer))
 			ft_putendl_fd("error: Could not allocate memory!", 2);
 		else
 			confirm = false;
 	}
 	else if (elem)
-		confirm = ft_rbtree_delete(tree, elem, &free) == NULL;
+		confirm = ft_rbtree_delete(tree, elem, NULL) == NULL;
 	else
 		ft_putendl_fd("error: Value not found!", 2);
 	if (confirm)
@@ -68,12 +59,12 @@ static void		edit_tree(t_rbtree **tree, char op)
 
 static void		show_tree(t_rbtree *tree)
 {
-	ft_rbtree_print(tree, &ft_print_int);
+	ft_rbtree_print(tree, &print_value);
 	ft_printf("size: %lu\n", ft_rbtree_size(tree));
 	if (tree)
 	{
-		ft_printf("min: %d\n", *((int *)(ft_rbtree_min(tree)->content)));
-		ft_printf("max: %d\n", *((int *)(ft_rbtree_max(tree)->content)));
+		ft_printf("min: %lu\n", (t_u64)(ft_rbtree_min(tree)->content));
+		ft_printf("max: %lu\n", (t_u64)(ft_rbtree_max(tree)->content));
 	}
 	ft_memdel(get_line(CONTINUE));
 }
@@ -83,15 +74,17 @@ int				main(void)
 	t_rbtree	*tree;
 	char		*s;
 	char		c;
+	int			r;
 
 	tree = NULL;
-	while (1)
+	r = 1;
+	while (r > 0)
 	{
 		ft_putstr("\033[1;1H\033[2JRED BLACK TREE - MENU\na: add a new value\n"
-		"d: delete a value\nv: tree view\nq: quit\n\n > ");
-		if (ft_get_next_line(0, &s) >= 0)
+		"d: delete a value\nv: tree view\nQuit with Ctrl+D\n\n > ");
+		if ((r = ft_get_next_line(0, &s)) >= 0)
 		{
-			c = *s && !s[1] ? *s : 0;
+			c = *s && !s[1] ? *s : '\0';
 			ft_memdel(s);
 			if (c == 'a' || c == 'd')
 				edit_tree(&tree, c);
@@ -101,6 +94,6 @@ int				main(void)
 				break ;
 		}
 	}
-	ft_rbtree_clear(&tree, &free);
+	ft_rbtree_clear(&tree, NULL);
 	return (0);
 }