+++ src/list/ft_list_pop.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/16 01:13:28 by asoursou          #+#    #+#             */
-/*   Updated: 2020/04/25 16:36:16 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/09/17 19:18:46 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,18 +15,18 @@
 
 void	*ft_list_pop(t_list **l, t_gfunction del)
 {
-	t_list	*next;
+	t_list	*elem;
 	void	*content;
 
-	next = (*l)->next;
+	if (!(elem = ft_list_popl(l)))
+		return (NULL);
 	if (del)
 	{
-		del((*l)->content);
+		del(elem->content);
 		content = NULL;
 	}
 	else
-		content = (*l)->content;
-	free(*l);
-	*l = next;
+		content = elem->content;
+	free(elem);
 	return (content);
 }