+++ src/list/ft_list_rev.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/06 00:33:17 by asoursou          #+#    #+#             */
-/*   Updated: 2020/04/12 18:24:59 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/09/17 19:13:42 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ t_list	*ft_list_rev(t_list *l)
 	while (l)
 	{
 		next = l->next;
-		ft_list_push(&prev, l);
+		ft_list_pushl(&prev, l);
 		l = next;
 	}
 	return (prev);