+++ src/list/ft_list_split.c
@@ -6,7 +6,7 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/27 15:50:44 by asoursou          #+#    #+#             */
-/*   Updated: 2020/09/10 12:31:41 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/09/17 19:12:52 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ t_list			*ft_list_split(const char *s, const char *set)
 	while ((n = ft_wordlen(&s, map)))
 	{
 		s2 = ft_memdup(s, n, n + 1, true);
-		if (!s2 || !ft_list_push(&l, ft_list_new(s2)))
+		if (!s2 || !ft_list_push(&l, s2))
 		{
 			s2 = ft_memdel(s2);
 			return (ft_list_clear(&l, &free));